+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 9.0.0-alpha.1
 
 ### Breaking Changes
 
@@ -49,6 +49,7 @@
 - Fix issue where the thread that generated an event could be missing when more than 100 threads are running (#6377)
 - Fix wrong Frame Delay when becoming active, which lead to false reported app hangs when the app moves to the foreground after being in the background (#6381)
 - Rename `SentryMechanismMeta` to `SentryMechanismContext` to resolve Kotlin Multi-Platform build errors (#6607)
+- Remove unnecesary dependency on `SentryCppHelper` to Sentry (#6754)
 
 ### Improvements
 
@@ -66,6 +67,7 @@
   This option is still disabled by default and will be enabled in a future major release.
 - Move `enableDataSwizzling` from experimental options to top-level options (#6592). This option remains enabled by default.
 - Add `sentry.replay_id` attribute to logs ([#6515](https://github.com/getsentry/sentry-cocoa/pull/6515))
+- Structured Logs: Add log APIs to `Hub` and `Client` (#6737)
 
 ## 9.0.0-alpha.0